+++ tokenizer_generator/include/finite_automaton/array/set_array.h
@@ -16,10 +16,10 @@ void kev_setarray_destroy(KevSetArray* array);
 bool kev_setarray_push_back(KevSetArray* array, KevBitSet* set);
 static inline void kev_setarray_pop_back(KevSetArray* array);
 bool kev_setarray_expand(KevSetArray* array);
-static inline KevBitSet* kev_setarray_visit(KevSetArray* array, uint64_t index);
-static inline uint64_t kev_setarray_size(KevSetArray* array);
+static inline KevBitSet* kev_setarray_visit(KevSetArray* array, size_t index);
+static inline size_t kev_setarray_size(KevSetArray* array);
 
-static inline uint64_t kev_setarray_size(KevSetArray* array) {
+static inline size_t kev_setarray_size(KevSetArray* array) {
   return array->current - array->begin;
 }
 
@@ -27,7 +27,7 @@ static inline void kev_setarray_pop_back(KevSetArray* array) {
   if (array) array->current--;
 }
 
-static inline KevBitSet* kev_setarray_visit(KevSetArray* array, uint64_t index) {
+static inline KevBitSet* kev_setarray_visit(KevSetArray* array, size_t index) {
   if (!array) return NULL;
   return array->begin[index];
 }
