+++ tokenizer_generator/src/finite_automaton/hashmap/intlist_map.c
@@ -4,23 +4,23 @@
 
 #include <stdlib.h>
 
-inline static uint64_t kev_intlistmap_hashing(uint64_t key) {
-  return (uint64_t)key;
+inline static size_t kev_intlistmap_hashing(size_t key) {
+  return (size_t)key;
 }
 
 static void kev_intlistmap_rehash(KevIntListMap* to, KevIntListMap* from) {
-  uint64_t from_capacity = from->capacity;
-  uint64_t to_capacity = to->capacity;
+  size_t from_capacity = from->capacity;
+  size_t to_capacity = to->capacity;
   KevIntListMapBucket* from_array = from->array;
   KevIntListMapBucket* to_array = to->array;
   KevIntListMapBucket* bucket_head = NULL;
-  uint64_t mask = to_capacity - 1;
-  for (uint64_t i = 0; i < from_capacity; ++i) {
+  size_t mask = to_capacity - 1;
+  for (size_t i = 0; i < from_capacity; ++i) {
     KevIntListMapNode* node = from_array[i].map_node_list;
     while (node) {
       KevIntListMapNode* tmp = node->next;
-      uint64_t hash_val = kev_intlistmap_hashing(node->key);
-      uint64_t index = hash_val & mask;
+      size_t hash_val = kev_intlistmap_hashing(node->key);
+      size_t index = hash_val & mask;
       node->next = to_array[index].map_node_list;
       to_array[index].map_node_list = node;
       if (node->next == NULL) {
@@ -59,15 +59,15 @@ static void kev_intlist_map_bucket_free(KevIntListMapBucket* bucket) {
   bucket->next = NULL;
 }
 
-inline static uint64_t pow_of_2_above(uint64_t num) {
-  uint64_t pow = 8;
+inline static size_t pow_of_2_above(size_t num) {
+  size_t pow = 8;
   while (pow < num)
     pow <<= 1;
 
   return pow;
 }
 
-bool kev_intlistmap_init(KevIntListMap* map, uint64_t capacity) {
+bool kev_intlistmap_init(KevIntListMap* map, size_t capacity) {
   if (!map) return false;
 
   map->bucket_head = NULL;
@@ -80,7 +80,7 @@ bool kev_intlistmap_init(KevIntListMap* map, uint64_t capacity) {
     return false;
   }
 
-  for (uint64_t i = 0; i < capacity; ++i) {
+  for (size_t i = 0; i < capacity; ++i) {
     array[i].map_node_list = NULL;
   }
   
@@ -106,14 +106,14 @@ void kev_intlistmap_destroy(KevIntListMap* map) {
   }
 }
 
-KevIntListMapNode* kev_intlistmap_insert(KevIntListMap* map, uint64_t key, KevNodeList* value) {
+KevIntListMapNode* kev_intlistmap_insert(KevIntListMap* map, size_t key, KevNodeList* value) {
   if (map->size >= map->capacity && !kev_intlistmap_expand(map))
     return NULL;
 
   KevIntListMapNode* new_node = kev_intlistmap_node_pool_allocate();
   if (!new_node) return NULL;
 
-  uint64_t index = (map->capacity - 1) & kev_intlistmap_hashing(key);
+  size_t index = (map->capacity - 1) & kev_intlistmap_hashing(key);
   new_node->key = key;
   new_node->value = value;
   new_node->next = map->array[index].map_node_list;
@@ -126,8 +126,8 @@ KevIntListMapNode* kev_intlistmap_insert(KevIntListMap* map, uint64_t key, KevNo
   return new_node;
 }
 
-KevIntListMapNode* kev_intlistmap_search(KevIntListMap* map, uint64_t key) {
-  uint64_t index = (map->capacity - 1) & kev_intlistmap_hashing(key);
+KevIntListMapNode* kev_intlistmap_search(KevIntListMap* map, size_t key) {
+  size_t index = (map->capacity - 1) & kev_intlistmap_hashing(key);
   KevIntListMapNode* node = map->array[index].map_node_list;
   while (node) {
     if (node->key == key)
@@ -151,7 +151,7 @@ void kev_intlistmap_make_empty(KevIntListMap* map) {
 
 KevIntListMapNode* kev_intlistmap_iterate_next(KevIntListMap* map, KevIntListMapNode* current) {
   if (current->next) return current->next;
-  uint64_t index = (map->capacity - 1) & kev_intlistmap_hashing(current->key);
+  size_t index = (map->capacity - 1) & kev_intlistmap_hashing(current->key);
   KevIntListMapBucket* current_bucket = &map->array[index];
   if (current_bucket->next)
     return current_bucket->next->map_node_list;
